requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.17.0
numpy>=1.24.0
python-dotenv>=1.0.0
openai>=1.3.0

# Google Sheets Integration
gspread>=5.12.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
openpyxl>=3.1.0
